Added docs, moved JsonMessage to own namespace.
LoginStruct/RegisterStruct:
- add: Documentation about the type, its usage and members.

JsonMessage:
- refac: Moved JsonMessage*-classes to own namespace to reduce clutter.
