+++ Shared/Json/JsonMessageClient.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace Shared
+﻿namespace Shared.Json
 {
     internal class JsonMessageClient
     {
