+++ Shared/Message.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Shared.Json;
+using System;
 using System.Drawing;
 using System.Text.Json;
